修改一个配置

+++ IceKuangjia/IceKuangjia/Lib/GlobalMethod/SLUChatToolBar.m
@@ -65,7 +65,7 @@
 }
 
 - (void)btnClick:(UIButton *)sender{
-    NSLog(@"%ld", sender.tag);
+    NSLog(@"%ld", (long)sender.tag);
 }
 
 - (void)show{
